requirements.txt
decorator==4.0.10
dj-database-url==0.4.1
Django==1.10.1
django-registration==2.1.2
djangorestframework==3.4.7
factory-boy==2.7.0
fake-factory==0.7.2
ipython==5.1.0
ipython-genutils==0.1.0
pexpect==4.2.1
pickleshare==0.7.4
Pillow==3.3.1
prompt-toolkit==1.0.7
psycopg2==2.6.2
ptyprocess==0.5.1
Pygments==2.1.3
python-dateutil==2.5.3
simplegeneric==0.8.1
six==1.10.0
traitlets==4.3.0
wcwidth==0.1.7
